Enhance FluentDocker with various improvements and fixes
- Updated ResourceQuery to clarify resource name extraction and improve assembly loading logic.
- Refactored ResourceReader to implement a custom enumerator for better resource stream handling.
- Improved ServiceEndpointResolver to ensure custom resolvers are consulted for portless containers.
- Enhanced ServiceExtensions to provide better error handling and diagnostics during process waits.
- Clarified documentation regarding volume deletion behavior on container removal.
- Modified ComposeService to ensure hooks are executed only after confirming the project is running.
- Improved ContainerService lifecycle management to handle state transitions more accurately.
- Updated HostService to better manage image pulling and conflict resolution with digest references.
- Enhanced ModelService to ensure proper disposal handling and state management.
- Improved TopologyResource to generate unique resource names for parallel deployments.
- Added internal AI-remediation prompt templates for production readiness checks.
- Updated README to clarify package changes and migration guidance for users transitioning from Ductus.FluentDocker 2.x.

+++ FluentDocker/Drivers/Docker/Api/ApiModels/DockerApiContainerModels.cs
@@ -206,7 +206,9 @@ namespace FluentDocker.Drivers.Docker.Api.ApiModels
   /// </summary>
   internal sealed class ExecInspectResponse
   {
-    [JsonPropertyName("ExitCode")] public int? ExitCode { get; set; }
+    // int64 in the Docker API contract: Windows containers exit with values like
+    // 3221225477 (0xC0000005) that overflow Int32 and would fail deserialization.
+    [JsonPropertyName("ExitCode")] public long? ExitCode { get; set; }
     [JsonPropertyName("Running")] public bool Running { get; set; }
   }
 
@@ -248,7 +250,8 @@ namespace FluentDocker.Drivers.Docker.Api.ApiModels
   /// </summary>
   internal sealed class WaitContainerResponse
   {
-    [JsonPropertyName("StatusCode")] public int StatusCode { get; set; }
+    // int64 in the Docker API contract (see ExecInspectResponse.ExitCode).
+    [JsonPropertyName("StatusCode")] public long StatusCode { get; set; }
 
     [JsonPropertyName("Error")]
     public WaitContainerError Error { get; set; }
